refactored the tests in unit versus regression
