+++ tests/test_regression_embedded_fonts.py
