+++ tests/test_regression_threaded_parse.py
@@ -22,7 +22,7 @@ from tests.constants import (
     PARSER_PAGE_RESTRICTIONS,
     SAMPLE_PDF,
 )
-from tests.test_parse import (
+from tests.test_regression_parse import (
     GROUNDTRUTH_FOLDER,
     REGRESSION_FOLDER,
     verify_SegmentedPdfPage,
