+++ tests/test_unit_ccitt_images.py
