+++ tests/test_unit_clipping.py
