+++ tests/test_unit_indexed_images.py
